Revert "Refactor asciiviz into modular sources"
